+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Migrate the whole build process to Webpack 5. This should have no impact on
+  end user but it's worth mentioning.
+
 ## v1.1.1 - 2020-08-18
 
 ### Fixed