+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Migrate the whole build process to Webpack 5. This should have no impact on
-  end user but it's worth mentioning.
+  end users but it's worth mentioning.
+- Migrate tests to Jest. This should have no impact on end users but it's worth
+  mentioning.
 
 ## v1.1.1 - 2020-08-18