+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - It is now possible to change the text for the "Top Banner" and "Bottom Banner"
   style warnings.
+- It is now possible to change the colors of all warnings.
 
 ### Changed